dom/tests/mochitest/bugs/test_bug585240.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=585240
-->
<head>
  <title>Test for Bug 585240</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=585240">Mozilla Bug 585240</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 585240 **/

  SimpleTest.waitForExplicitFinish();

  window.onmessage = function(event) {
    ok(true, "message event should fire!");
    SimpleTest.finish();
  }
  window.postMessage("hello", "*");  

</script>
</pre>
</body>
</html>
